refactor: Move mustUnmarshalYaml to files.go
